api/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pymongo[srv]==4.6.1
python-dotenv==1.0.0
bcrypt>=4.0.0
email-validator==2.1.0
